fixed logdata configuration file
